Added web crawling classes
The classes that were added are essential to web crawling operations.
There is now a user agent assigner that loads user agents from a file,
a CrawlerManager that will give the user information about the crawling
operations and manages links, and the Crawler that scans the web pages
and relays its results to the CrawlerManager.

There were a few modifications to the MainGraphics class. A few field 
variables will be added, but the MainGraphics will be modified when the
web crawler works. In other words, they will be the last thing to work
on.