Web crawling is working but needs tune-ups and testing.
The CrawlerLeg class was added to scan the page for words and gather 
links. The InformationPackage contains information about what crawler
legs found on a page. The Crawler creates the CrawlerLegs. The
CrawlerManager manages every Crawler. This makes it so the GUI can better
obtain information about scanning activities. The QuickWrite was added 
to write what the program was doing to a file. The statements are 
commented out though so nothing will be written to a log file. Look at the
log file under the "src" folder to get a perspective on what the program
is doing. This is a multi threaded web crawling program that can crawl
on multiple links at one time. Run the "WebCrawlUtilitiesTester" class
to start the web crawler for now. After some problems are solved, then
the GUI will start to be designed. Also, the agents.txt file contains 
user agents. The UserAgentAssigner loads a specified amount of user agents
from the file and assigns them at random.