Added some functionality to the start button.
The start button will now open a dialog box that allows the user to 
specify starting sites to crawl, the amount of sites to crawl, and some
search queries to look for while crawling. There is no functionality in
this dialog box though, except for the fact that you can add websites 
to the sites list.