You can now use the GUI for web crawling.
Valid links are now added to the valid links JList when crawling. The 
crawler setup dialog is now sort of finished, needs a little bit of 
improvements. A lot of updates will be made in the coming days. Also, 
commenting was not updated for this commit and the commenting will be 
updated on the next commit (tomorrow).

View the